Update Assignment2.java
Changed getInt('grade') to getFloat('grade') on query7 line 418.

+++ Assignment2.java
@@ -415,7 +415,7 @@ public class Assignment2 {
                                 rs.getString("cname").trim() + ":" +
                                 rs.getInt("year") + ":" +
                                 rs.getInt("semester") + ":" +
-                                rs.getInt("grade");
+                                rs.getFloat("grade");
                 }
                 // Close the resultset
                 rs.close();
